+++ tools/mobileExperience/tool.js
@@ -83,7 +83,8 @@
       const parent = this.lichessTools;
       const lichess = parent.lichess;
       const $ = parent.$;
-      const pos = [e.targetTouches[0].clientX, e.targetTouches[0].clientY];
+      const ev = e.targetTouches?.[0] || e;
+      const pos = [ev.clientX, ev.clientY];
       const square = this.chessground.getKeyAtDomPos(pos);
       this.chessground.state.drawable.current = {
         orig: square,
@@ -99,7 +100,8 @@
       const lichess = parent.lichess;
       const $ = parent.$;
       if (!this.chessground.state.drawable.current) return;
-      const pos = [e.targetTouches[0].clientX, e.targetTouches[0].clientY];
+      const ev = e.targetTouches?.[0] || e;
+      const pos = [ev.clientX, ev.clientY];
       const square = this.chessground.getKeyAtDomPos(pos);
       const current = this.chessground.state.drawable.current;
       current.pos = pos;
@@ -115,7 +117,8 @@
       if (!this.chessground.state.drawable.current) return;
       e.preventDefault();
       e.stopPropagation();
-      const pos = [e.changedTouches[0].clientX, e.changedTouches[0].clientY];
+      const ev = e.targetTouches?.[0] || e;
+      const pos = [ev.clientX, ev.clientY];
       const square = this.chessground.getKeyAtDomPos(pos);
       this.handleGesture(this.chessground.state.drawable.current);
       this.chessground.state.drawable.current = undefined;
@@ -179,6 +182,7 @@
     brushes = ['green', 'red', 'blue', 'yellow'];
     toggleBrush = (ev) => {
       if (!this.chessground) return;
+      ev.preventDefault();
       let index = this.brushes.indexOf(this.drawingBrush) + 1;
       this.drawingBrush = index >= this.options.colorCount
         ? null
@@ -190,6 +194,19 @@
       for (const brush of this.brushes) {
         $(ev.target)
           .toggleClass('lichessTools-' + brush + 'Brush', this.drawingBrush == brush);
+      }
+    };
+
+    clickOrTapAnalysisControls = (ev) => {
+      const parent = this.lichessTools;
+      const $ = parent.$;
+      if ($(ev.target).is('button.lichessTools-shapeDrawing')) {
+        ev.preventDefault();
+        this.toggleBrush(ev);
+      }
+      if ($(ev.target).is('button.lichessTools-randomNextMove')) {
+        ev.preventDefault();
+        this.playRandomVariation();
       }
     };
 
@@ -231,17 +248,17 @@
         if (wrap && !wrap.is('.lichessTools-shapeDrawing')) {
           wrap
             .addClass('lichessTools-shapeDrawing')
-            .on('touchstart', this.touchStart)
-            .on('touchmove', this.touchMove)
-            .on('touchend', this.touchEnd);
+            .on('touchstart mousedown ', this.touchStart)
+            .on('touchmove mousemove', this.touchMove)
+            .on('touchend mouseup', this.touchEnd);
         }
       } else {
         if (wrap) {
           wrap
             .removeClass('lichessTools-shapeDrawing')
-            .off('touchstart', this.touchStart)
-            .off('touchmove', this.touchMove)
-            .off('touchend', this.touchEnd);
+            .off('touchstart mousedown', this.touchStart)
+            .off('touchmove mousemove', this.touchMove)
+            .off('touchend mouseup', this.touchEnd);
         }
       }
       if (isAnalyse) {
@@ -279,17 +296,12 @@
               this.originalHandler = parent.getEventHandlers(elem, 'touchstart')?.at(0)?.bind(elem);
             }
             parent.removeEventHandlers(elem, 'touchstart');
-            $('div.analyse__controls').on('touchstart', ev => {
-              this.originalHandler(ev);
-              if ($(ev.target).is('button.lichessTools-shapeDrawing')) {
-                ev.preventDefault();
-                this.toggleBrush(ev);
-              }
-              if ($(ev.target).is('button.lichessTools-randomNextMove')) {
-                ev.preventDefault();
-                this.playRandomVariation();
-              }
-            });
+            $('div.analyse__controls')
+              .on('touchstart', ev => {
+                this.originalHandler(ev);
+                this.clickOrTapAnalysisControls(ev);
+              })
+              .on('mousedown', this.clickOrTapAnalysisControls);
           }
         }
         if (!this.options.shapeDrawing && !this.options.randomNextMove) {
@@ -297,7 +309,9 @@
             const elem = $('.analyse__controls')[0];
             if (elem && this.originalHandler) {
               parent.removeEventHandlers(elem, 'touchstart');
-              $('div.analyse__controls').on('touchstart', this.originalHandler);
+              $('div.analyse__controls')
+                .on('touchstart', this.originalHandler)
+                .off('mousedown', this.clickOrTapAnalysisControls);
             }
           }
         }
@@ -310,7 +324,7 @@
                 .attr('data-icon', '\u21D7')
                 .attr('title', trans.noarg('shapeDrawingTitle'))
                 .insertBefore($('button.board-menu-toggle', container))
-                .on('touchstart', ev => {
+                .on('touchstart mousedown ', ev => {
                   this.toggleBrush(ev);
                   wrap?.toggleClass('lichessTools-passthrough', !this.drawingBrush);
                 });